update locations data structure
SteamAPI::GetLocalPingLocation return the age of the data and edit (by reference)
the result SteamNetworkPingLocation_t which contains the location.

From Steam point of view, a location is an array of 512 bytes-length.
This is an opaque item that can be changed to String to be sent over network and
deserialized by after.

The commit changes all methods that uses location to take the location structure:

getLocalPingLocation
estimatePingTimeBetweenTwoLocations
estimatePingTimeFromLocalHost
convertPingLocationToString
parsePingLocationString